# Fix balances query performance with related schema changes