**Detailed description**:

The balances query on account_balances was doing a table scan because the account_id, as a user defined type `entity_id` couldn't be indexed by postgres' btree indexer.

It can be, if additional operators and an operator class are defined, but that requires superuser privilege.

So - the entity_id type is deprecated, and in its place account_balances has account_realm_num and account_num columns.

- Database upgrade script will migrate _all_ data from the former account_id field to account_realm_num and account_num columns
- Change deprecated table and column names to suffix __deprecated (remove them in next release via #175 )
- Update balances.js query to use new columns.

**Which issue(s) this PR fixes**:
Fixes #173

**Special notes for your reviewer**:

I manually tested the upgrade from the prior database version with 2.5M rows already in account balances.

I ran a local performance test of inserts and query before and after the upgrade.
Before:
- insert 1M rows - 25sec
- query against 2.5M rows - 0.5sec
After:
- query against 2.5M rows : 0.5ms (milliseconds)
- clear DB and insert 1M rows again : 19sec

here is the query plan after upgrade (see that it uses index):
```
explain analyze select consensus_timestamp, account_realm_num, account_num from account_balances where account_realm_num = 0 and account_num = 0 order by consensus_timestamp desc, account_realm_num desc, account_num desc;

 Sort  (cost=493.03..493.34 rows=125 width=51) (actual time=0.397..0.399 rows=25 loops=1)
   Sort Key: consensus_timestamp DESC
   Sort Method: quicksort  Memory: 28kB
   ->  Bitmap Heap Scan on account_balances  (cost=5.71..488.68 rows=125 width=51) (actual time=0.088..0.350 rows=25 loops=1)
         Recheck Cond: (((account_realm_num)::smallint = 0) AND ((account_num)::integer = 0))
         Heap Blocks: exact=25
         ->  Bitmap Index Scan on idx__account_balances__account_then_timestamp  (cost=0.00..5.68 rows=125 width=0) (actual time=0.072..0.072 rows=25 loops=1)
               Index Cond: (((account_realm_num)::smallint = 0) AND ((account_num)::integer = 0))
 Planning time: 0.778 ms
 Execution time: 0.470 ms
```

**Checklist**
- [ ] Documentation added
- [ ] Tests updated

Tests not updated as still awaiting integration/performance tooling setup.